backend/requirements.txt
alembic
flask
flask-cors
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
python-dotenv
werkzeug
wtforms